Add: Logging & Suspend & Elevate users

+++ API/API/Controllers/GameController.cs
@@ -1,7 +1,9 @@
 ﻿using API.Data;
 using API.Models;
+using Azure.Core;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System;
 
 namespace API.Controllers
 {
@@ -20,18 +22,30 @@ namespace API.Controllers
         [HttpGet("{token}")]
         public async Task<ActionResult<Status>> StatusByToken(string token)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(token, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                    new(name, "FAIL:Game/StatusByToken/Check", $"Failed to pass the check before game status from player {token} was fetched out of the game database within the game controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.GameRepository.GetStatusByPlayersToken(token);
 
             if (response is null)
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Game/StatusByToken", $"Failed to fetch game status from game {token} out of the game database within the game controller.")
+                    new(name, "FAIL:Game/StatusByToken", $"Failed to fetch game status from player {token} out of the game database within the game controller.")
                 );
                 return NotFound();
             }
 
             await _repository.LogRepository.Create(
-                new(User?.Identity?.Name ?? "Anonymous", "Game/StatusByToken", $"Fetched game status from game {token} with status {response} from the game database within the game controller.")
+                new(name, "Game/StatusByToken", $"Fetched game status from player {token} with status {response} from the game database within the game controller.")
             );
 
             return Ok(response);
@@ -40,18 +54,30 @@ namespace API.Controllers
         [HttpGet("view/{token}")]
         public async Task<ActionResult<GamePlay>> View(string token)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(token, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                    new(name, "FAIL:Game/GetView/Check", $"Failed to pass the check before game view from player {token} was fetched out of the game database within the game controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.GameRepository.GetView(token);
 
             if (response is null)
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Game/GetView", $"Failed to fetch game view of player {token} from game database within the game controller.")
+                    new(name, "FAIL:Game/GetView", $"Failed to fetch game view of player {token} from game database within the game controller.")
                 );
                 return NotFound();
             }
 
             await _repository.LogRepository.Create(
-                new(User?.Identity?.Name ?? "Anonymous", "Game/GetView", $"Fetch game view from player {token} from the game database within the game controller.")
+                new(name, "Game/GetView", $"Fetch game view from player {token} from the game database within the game controller.")
             );
 
             return Ok(response);
@@ -60,18 +86,30 @@ namespace API.Controllers
         [HttpGet("partial/{token}")]
         public async Task<ActionResult<GamePartial>> Partial(string token)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(token, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                    new(name, "FAIL:Game/Partial/Check", $"Failed to pass the check before game partial from player {token} was fetched out of the game database within the game controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.GameRepository.GetPartial(token);
 
             if (response is null)
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Game/GetPartial", $"Failed to fetch game partial of player {token} from game database within the game controller.")
+                    new(name, "FAIL:Game/GetPartial", $"Failed to fetch game partial of player {token} from game database within the game controller.")
                 );
                 return NotFound();
             }
 
             await _repository.LogRepository.Create(
-                new(User?.Identity?.Name ?? "Anonymous", "Game/GetPartial", $"Fetch game partial from player {token} from game database within the game controller.")
+                new(name, "Game/GetPartial", $"Fetch game partial from player {token} from game database within the game controller.")
             );
 
             return Ok(response);
@@ -80,20 +118,32 @@ namespace API.Controllers
         [HttpPost("create")]
         public async Task<ActionResult<HttpResponseMessage>> Create([FromBody] GameCreation game)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(game.PlayerToken, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                    new(name, "FAIL:Game/Create/Check", $"Failed to pass the check before game partial from player {game.PlayerToken} was fetched out of the game database within the game controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.GameRepository.Create(game);
 
             if (response == true)
             {
                 await _repository.PlayerRepository.UpdateActivity(game.PlayerToken);
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Game/Create", $"Player {game.PlayerToken} created game within the game controller.")
+                    new(name, "Game/Create", $"Player {game.PlayerToken} created game within the game controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Game/Create", $"Player {game.PlayerToken} failed to creat game within the game controller.")
+                    new(name, "FAIL:Game/Create", $"Player {game.PlayerToken} failed to creat game within the game controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
             }
@@ -102,6 +152,18 @@ namespace API.Controllers
         [HttpPost("join")]
         public async Task<ActionResult<HttpResponseMessage>> Join([FromBody] PlayerRequest request)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(request.SenderToken, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                    new(name, "FAIL:Game/Join/Check", $"Failed to pass the check before player {request.SenderToken} joinded player's {request.ReceiverUsername} game from within the game controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.GameRepository.JoinPlayer(request);
 
             if (response == true)
@@ -113,13 +175,13 @@ namespace API.Controllers
                     await _repository.PlayerRepository.UpdateActivity(player.Token);
                     await _repository.PlayerRepository.UpdateActivity(request.SenderToken);
                     await _repository.LogRepository.Create(
-                        new(User?.Identity?.Name ?? "Anonymous", "Game/Join", $"Player {request.SenderToken} joined {request.ReceiverUsername}'s game from within the game controller.")
+                        new(name, "Game/Join", $"Player {request.SenderToken} joined {request.ReceiverUsername}'s game from within the game controller.")
                     );
                     return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
                 }
             }
             await _repository.LogRepository.Create(
-                new(User?.Identity?.Name ?? "Anonymous", "FAIL:Game/Join", $"Player {request.SenderToken} failed to join {request.ReceiverUsername}'s game from within the game controller.")
+                new(name, "FAIL:Game/Join", $"Player {request.SenderToken} failed to join {request.ReceiverUsername}'s game from within the game controller.")
             );
             return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
         }
@@ -127,20 +189,32 @@ namespace API.Controllers
         [HttpPost("move")]
         public async Task<ActionResult<HttpResponseMessage>> Move([FromBody] GameMove action)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(action.PlayerToken, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Game/Move/Check", $"Failed to pass the check before player {action.PlayerToken} made a move from within the game controller.")
+                );
+                return BadRequest();
+            }
+
             var (succeded, error) = await _repository.GameRepository.Move(action);
 
             if (succeded)
             {
                 await _repository.PlayerRepository.UpdateActivity(action.PlayerToken);
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Game/Move", $"Player {action.PlayerToken} made move ({action.Row},{action.Column}) from within game controller.")
+                    new(name ?? "Anonymous", "Game/Move", $"Player {action.PlayerToken} made move ({action.Row},{action.Column}) from within game controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Game/Move", $"Player {action.PlayerToken} failed to make move ({action.Row},{action.Column}) from within the game controller.")
+                    new(name ?? "Anonymous", "FAIL:Game/Move", $"Player {action.PlayerToken} failed to make move ({action.Row},{action.Column}) from within the game controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest)
                 {
@@ -152,20 +226,32 @@ namespace API.Controllers
         [HttpPost("pass")]
         public async Task<ActionResult<HttpResponseMessage>> Pass([FromBody] ID id)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(id.Token, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Game/Pass/Check", $"Failed to pass the check before player {id.Token} passed from within the game controller.")
+                );
+                return BadRequest();
+            }
+
             var (succeded, error) = await _repository.GameRepository.Pass(id.Token);
 
             if (succeded)
             {
                 await _repository.PlayerRepository.UpdateActivity(id.Token);
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Game/Pass", $"Player {id.Token} passed his turn from within the game controller.")
+                    new(name, "Game/Pass", $"Player {id.Token} passed his turn from within the game controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Game/Pass", $"Player {id.Token} passed his turn from within the game controller.")
+                    new(name, "FAIL:Game/Pass", $"Player {id.Token} passed his turn from within the game controller.")
                 ); 
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest)
                 {
@@ -177,20 +263,32 @@ namespace API.Controllers
         [HttpPost("forfeit")]
         public async Task<ActionResult<HttpResponseMessage>> Forfeit([FromBody] ID id)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(id.Token, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Game/Forfeit/Check", $"Failed to pass the check before player {id.Token} forfeited from within the game controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.GameRepository.Forfeit(id.Token);
 
             if (response == true)
             {
                 await _repository.PlayerRepository.UpdateActivity(id.Token);
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Game/Forfeit", $"Player {id.Token} forfeited the game from within the game controller.")
+                    new(name, "Game/Forfeit", $"Player {id.Token} forfeited the game from within the game controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Game/Forfeit", $"Player {id.Token} failed to forfeit the game from within the game controller.")
+                    new(name, "FAIL:Game/Forfeit", $"Player {id.Token} failed to forfeit the game from within the game controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
             }
@@ -199,20 +297,32 @@ namespace API.Controllers
         [HttpPost("delete")]
         public async Task<ActionResult<HttpResponseMessage>> Delete([FromBody] ID id)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(id.Token, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Game/Delete/Check", $"Failed to pass the check before player {id.Token} was deleted from within the game controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.GameRepository.Delete(id.Token);
 
             if (response == true)
             {
                 await _repository.PlayerRepository.UpdateActivity(id.Token);
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Game/Delete", $"Player {id.Token} deleted the game from within the game controller.")
+                    new(name, "Game/Delete", $"Player {id.Token} deleted the game from within the game controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Game/Delete", $"Player {id.Token} failed to delete the game from within the game controller.")
+                    new(name, "FAIL:Game/Delete", $"Player {id.Token} failed to delete the game from within the game controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
             }
