+++ API/API/Controllers/PlayerController.cs
@@ -1,5 +1,6 @@
 ﻿using API.Data;
 using API.Models;
+using Azure.Core;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -21,52 +22,56 @@ namespace API.Controllers
         public async Task<ActionResult<string?>> Rematch(string token)
         {
             string[] parts = token.Split(' ');
+
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(parts[1], name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Player/Rematch/Check", $"Failed to pass the check before player {parts[1]} fetched the rematch data from player {parts[0]} within the player controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.PlayerRepository.GetRematch(parts[0], parts[1]);
 
             await _repository.LogRepository.Create(
-                new(User?.Identity?.Name ?? "Anonymous", "Player/Rematch", $"Player {parts[1]} {(response == null ? "didn't found" : "did found")} a rematch with player {parts[0]} within the player controller.")
+                new(name, "Player/Rematch", $"Player {parts[1]} {(response == null ? "didn't found" : "did found")} a rematch with player {parts[0]} within the player controller.")
             );
 
             return Ok(response);
         }
 
-        [HttpPost("activity")]
-        public async Task<ActionResult<HttpResponseMessage>> Activity([FromBody] ID id)
-        {
-            var response = await _repository.PlayerRepository.UpdateActivity(id.Token);
-
-            if (response == true)
-            {
-                await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Player/Activity", $"Updated last activity from player {id.Token} within the player controller.")
-                );
-                return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
-            }
-            else
-            {
-                await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Player/Activity", $"Failed to update last activity from player {id.Token} within the player controller.")
-                );
-                return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
-            }
-        }
-
         [HttpPost("request/friend")]
         public async Task<ActionResult<HttpResponseMessage>> FriendRequest([FromBody] PlayerRequest request)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(request.SenderToken, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Player/FriendRequest/Check", $"Failed to pass the check before player {request.SenderToken} sent a friend request to player {request.ReceiverUsername} within the player controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.PlayerRepository.FriendRequest(request.ReceiverUsername, request.SenderToken);
 
             if (response == true)
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Player/FriendRequest", $"Player {request.SenderToken} sent a friend request to player {request.ReceiverUsername} within the player controller.")
+                    new(name, "Player/FriendRequest", $"Player {request.SenderToken} sent a friend request to player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Player/FriendRequest", $"Player {request.SenderToken} failed to send a friend request to player {request.ReceiverUsername} within the player controller.")
+                    new(name, "FAIL:Player/FriendRequest", $"Player {request.SenderToken} failed to send a friend request to player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
             }
@@ -75,19 +80,31 @@ namespace API.Controllers
         [HttpPost("request/friend/accept")]
         public async Task<ActionResult<HttpResponseMessage>> AcceptFriendRequest([FromBody] PlayerRequest request)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(request.SenderToken, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Player/AcceptFriend/Check", $"Failed to pass the check before player {request.SenderToken} accepted the friend request from player {request.ReceiverUsername} within the player controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.PlayerRepository.AcceptFriendRequest(request.ReceiverUsername, request.SenderToken);
 
             if (response == true)
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Player/AcceptFriendRequest", $"Player {request.SenderToken} accepted the a friend request of player {request.ReceiverUsername} within the player controller.")
+                    new(name, "Player/AcceptFriendRequest", $"Player {request.SenderToken} accepted the a friend request of player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Player/AcceptFriendRequest", $"Player {request.SenderToken} failed to accept the a friend request of player {request.ReceiverUsername} within the player controller.")
+                    new(name, "FAIL:Player/AcceptFriendRequest", $"Player {request.SenderToken} failed to accept the a friend request of player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
             }
@@ -96,19 +113,31 @@ namespace API.Controllers
         [HttpPost("request/friend/decline")]
         public async Task<ActionResult<HttpResponseMessage>> DeclineFriendRequest([FromBody] PlayerRequest request)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(request.SenderToken, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Player/DeclineFriendRequest/Check", $"Failed to pass the check before player {request.SenderToken} declined the friend request from player {request.ReceiverUsername} within the player controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.PlayerRepository.DeclineFriendRequest(request.ReceiverUsername, request.SenderToken);
 
             if (response == true)
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Player/DeclineFriendRequest", $"Player {request.SenderToken} declined the a friend request of player {request.ReceiverUsername} within the player controller.")
+                    new(name, "Player/DeclineFriendRequest", $"Player {request.SenderToken} declined the a friend request of player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Player/DeclineFriendRequest", $"Player {request.SenderToken} failed to decline the a friend request of player {request.ReceiverUsername} within the player controller.")
+                    new(name, "FAIL:Player/DeclineFriendRequest", $"Player {request.SenderToken} failed to decline the a friend request of player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
             }
@@ -117,19 +146,31 @@ namespace API.Controllers
         [HttpPost("request/game")]
         public async Task<ActionResult<HttpResponseMessage>> GameRequest([FromBody] PlayerRequest request)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(request.SenderToken, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Player/GameRequest/Check", $"Failed to pass the check before player {request.SenderToken} sent a game request to player {request.ReceiverUsername} within the player controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.PlayerRepository.GameRequest(request.ReceiverUsername, request.SenderToken);
 
             if (response == true)
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Player/GameRequest", $"Player {request.SenderToken} sent a game request to player {request.ReceiverUsername} within the player controller.")
+                    new(name, "Player/GameRequest", $"Player {request.SenderToken} sent a game request to player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Player/GameRequest", $"Player {request.SenderToken} failed to send a game request to player {request.ReceiverUsername} within the player controller.")
+                    new(name, "FAIL:Player/GameRequest", $"Player {request.SenderToken} failed to send a game request to player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
             }
@@ -138,19 +179,31 @@ namespace API.Controllers
         [HttpPost("request/game/accept")]
         public async Task<ActionResult<HttpResponseMessage>> AcceptGameRequest([FromBody] PlayerRequest request)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(request.SenderToken, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Player/AcceptGameRequest/Check", $"Failed to pass the check before player {request.SenderToken} accepted the game request from player {request.ReceiverUsername} within the player controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.PlayerRepository.AcceptGameRequest(request.ReceiverUsername, request.SenderToken);
 
             if (response == true)
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Player/AcceptGameRequest", $"Player {request.SenderToken} accepted the game request of player {request.ReceiverUsername} within the player controller.")
+                    new(name, "Player/AcceptGameRequest", $"Player {request.SenderToken} accepted the game request of player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Player/AcceptGameRequest", $"Player {request.SenderToken} failed to accept the game request of player {request.ReceiverUsername} within the player controller.")
+                    new(name, "FAIL:Player/AcceptGameRequest", $"Player {request.SenderToken} failed to accept the game request of player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
             }
@@ -159,19 +212,31 @@ namespace API.Controllers
         [HttpPost("request/game/decline")]
         public async Task<ActionResult<HttpResponseMessage>> DeclineGameRequest([FromBody] PlayerRequest request)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(request.SenderToken, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Player/DeclineGameRequest/Check", $"Failed to pass the check before player {request.SenderToken} declined the game request from player {request.ReceiverUsername} within the player controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.PlayerRepository.DeclineGameRequest(request.ReceiverUsername, request.SenderToken);
 
             if (response == true)
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Player/DeclineGameRequest", $"Player {request.SenderToken} declined the game request of player {request.ReceiverUsername} within the player controller.")
+                    new(name, "Player/DeclineGameRequest", $"Player {request.SenderToken} declined the game request of player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Player/DeclineGameRequest", $"Player {request.SenderToken} failed to decline the game request of player {request.ReceiverUsername} within the player controller.")
+                    new(name, "FAIL:Player/DeclineGameRequest", $"Player {request.SenderToken} failed to decline the game request of player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
             }
@@ -180,19 +245,31 @@ namespace API.Controllers
         [HttpPost("friend/delete")]
         public async Task<ActionResult<HttpResponseMessage>> DeleteFriend([FromBody] PlayerRequest request)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(request.SenderToken, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Player/DeleteFriend/Check", $"Failed to pass the check before player {request.SenderToken} deleted the friendship with player {request.ReceiverUsername} within the player controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.PlayerRepository.DeleteFriend(request.ReceiverUsername, request.SenderToken);
 
             if (response == true)
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "Player/DeleteFriend", $"Player {request.SenderToken} deleted the friendship with player {request.ReceiverUsername} within the player controller.")
+                    new(name, "Player/DeleteFriend", $"Player {request.SenderToken} deleted the friendship with player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
             }
             else
             {
                 await _repository.LogRepository.Create(
-                    new(User?.Identity?.Name ?? "Anonymous", "FAIL:Player/DeleteFriend", $"Player {request.SenderToken} failed to delete the friendship with player {request.ReceiverUsername} within the player controller.")
+                    new(name, "FAIL:Player/DeleteFriend", $"Player {request.SenderToken} failed to delete the friendship with player {request.ReceiverUsername} within the player controller.")
                 );
                 return new HttpResponseMessage(System.Net.HttpStatusCode.BadRequest);
             }
@@ -201,6 +278,18 @@ namespace API.Controllers
         [HttpPost("delete")]
         public async Task<ActionResult<HttpResponseMessage>> Delete([FromBody] ID id)
         {
+            var name = User?.Identity?.Name ?? "Anonymous";
+
+            var check = await _repository.PlayerRepository.PlayerChecksOut(id.Token, name);
+
+            if (check == false)
+            {
+                await _repository.LogRepository.Create(
+                new(name, "FAIL:Player/Delete/Check", $"Failed to pass the check before player {id.Token} deleted the account within the player controller.")
+                );
+                return BadRequest();
+            }
+
             var response = await _repository.PlayerRepository.Delete(id.Token);
 
             if (response == true)
